Positive Negative Pair
Given an array arr[] of N positive and negative integer pairs, may not be in sorted order. The task is to pair the positive and negative element in such a way that a positive element is paired with a negative element of same absolute value. If a pair element is not present for an element, then ignore it. The output should contain all pairs in increasing order of absolute values. To print a pair, positive value should be printed before its corresponding negative value.